requirements.txt
fastapi==0.108.0
fastui==0.4.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
python-dotenv==1.0.0
uvicorn==0.25.0